BK64: Move textures into their own module
